Separated app into frontend and backend
