+++ app/docker-compose.yml
@@ -1,0 +1,12 @@
+version: "3.8"
+services:
+  frontend:
+    build: ./frontend
+    ports:
+      - "5001:5001"
+    depends_on:
+      - backend
+  backend:
+    build: ./backend
+    ports:
+      - "5000:5000"
