+++ app/frontend/Dockerfile
@@ -10,11 +10,8 @@ ADD . /app
 # Install any needed packages specified in requirements.txt
 RUN pip install --no-cache-dir -r requirements.txt
 
-# Make port 5000 available to the world outside this container
-EXPOSE 5000
-
-# Define environment variable
-ENV NAME DockerApp
+# Make port 80 available to the world outside this container
+EXPOSE 5001
 
 # Run app.py when the container launches
 CMD ["python", "app.py"]
