+++ app/frontend/requirements.txt
@@ -1,0 +1,3 @@
+Flask==2.3.2
+Jinja2==3.1.2
+requests==2.31.0
