+++ app/frontend/templates/form.html
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+    <title>Supply Chain Simulator</title>
+</head>
+
+<body>
+    <h1>Supply Chain Simulator</h1>
+    <form action="/" method="post">
+        <!-- You could put form fields here for the user to specify parameters for the simulation. -->
+        <input type="submit" value="Run Simulation">
+    </form>
+</body>
+
+</html>
