+++ app/frontend/templates/result.html
@@ -1,0 +1,18 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+    <title>Simulation Result</title>
+</head>
+
+<body>
+    <h1>Simulation Result</h1>
+    <!-- The result of the simulation will be inserted here. -->
+    <p>The result is: {{ result }}</p>
+    <!-- Add a button to go back to the form. -->
+    <form action="/" method="get">
+        <input type="submit" value="Back">
+    </form>
+</body>
+
+</html>
